K8s: Split out the factory method.
This lets us import the K8s interfaces without pulling in the
implementations.
